Add README.md file
README is the most important file within the project. It usually is,
and it should be the first file to be opened and read before diving into
anything else. The file should be written for readability and
maintainability.

I was contemplating whether to choose reStructuredText or Markdown as
the markup language. As a Python advocate, I would rather use
reStructuredText as it has its roots stem from the programming
language's community. But on the other hand, despite having much less
feature, Markdown has much simpler syntax and is more ubiquitous
comparing to its counterpart.

As the result, I choose Markdown, as a better fitting tool for the job,
which is to write a simple README file.

P/s: The .md extension is for GitHub (and any future hosting service
provider) to properly display the file. Without the .md extension,
GitHub only sees it as plain text file rather than Markdown.